The examples are unified on one side
