+++ .vscode/launch.json
@@ -21,22 +21,6 @@
             "internalConsoleOptions": "openOnSessionStart"
         },
         {
-            "type": "node",
-            "request": "launch",
-            "name": "Tests samples2",
-            "program": "${workspaceFolder}/node_modules/mocha/bin/_mocha",
-            "args": [
-                "-u",
-                "tdd",
-                "--timeout",
-                "999999",
-                "--colors",
-                "${workspaceFolder}/dist/samples2/tests"
-            ],
-            "preLaunchTask": "compile",
-            "internalConsoleOptions": "openOnSessionStart"
-        },
-        {
             "args": [],
             "cwd": "${workspaceRoot}",
             "externalConsole": false,
@@ -52,23 +36,6 @@
             "sourceMaps": true,
             "stopOnEntry": false,
             "type": "node"
-        },
-        {
-            "args": [],
-            "cwd": "${workspaceRoot}",
-            "externalConsole": false,
-            "name": "DEBUG samples 2",
-            "outDir": "${workspaceRoot}/dist",
-            "program": "${workspaceRoot}/dist/samples2/server.js",
-            "request": "launch",
-            "runtimeArgs": [
-                "--nolazy"
-            ],
-            "preLaunchTask": "compile",
-            "runtimeExecutable": null,
-            "sourceMaps": true,
-            "stopOnEntry": false,
-            "type": "node"
-        },
+        }
     ]
 }
