--- samples2/environments/environment.prod.ts
@@ -1,4 +1,0 @@
-
-export const environment = {
-    name: 'prod'
-}
