--- samples2/environments/environment.ts
@@ -1,4 +1,0 @@
-export const environment = {
-    name: 'local',
-    domains: ['http://localhosts:4200', 'http://localhosts:4201']
-}
