--- samples2/utils.ts
@@ -1,7 +1,0 @@
-import { UserModel } from "./models/models";
-
-export class Utils {
-  public print() {
-    //console.log('entre1');
-  }
-}
