+++ tsconfig.json
@@ -14,7 +14,6 @@
   },
   "include": [
     "samples/**/*",
-    "samples2/**/*",
     "tests/**/*"
   ],
   "exclude": [
